+++ tests/sequelize/index.test.mjs
@@ -1,6 +1,6 @@
 import { Sequelize, DataTypes, Model } from 'sequelize';
 import pgvector from 'pgvector/sequelize';
-import { l2Distance, maxInnerProduct, cosineDistance, l1Distance } from 'pgvector/sequelize';
+import { l2Distance, maxInnerProduct, cosineDistance, l1Distance, hammingDistance, jaccardDistance } from 'pgvector/sequelize';
 
 test('example', async () => {
   pgvector.registerType(Sequelize);
@@ -19,6 +19,9 @@ test('example', async () => {
   const Item = sequelize.define('Item', {
     embedding: {
       type: DataTypes.VECTOR(3)
+    },
+    binary_embedding: {
+      type: 'BIT(3)'
     }
   }, {
     modelName: 'Item',
@@ -34,9 +37,9 @@ test('example', async () => {
 
   await Item.sync({force: true});
 
-  await Item.create({embedding: [1, 1, 1]});
-  await Item.create({embedding: [2, 2, 2]});
-  await Item.create({embedding: [1, 1, 2]});
+  await Item.create({embedding: [1, 1, 1], binary_embedding: '000'});
+  await Item.create({embedding: [2, 2, 2], binary_embedding: '101'});
+  await Item.create({embedding: [1, 1, 2], binary_embedding: '111'});
 
   // L2 distance
   let items = await Item.findAll({
@@ -70,6 +73,20 @@ test('example', async () => {
     limit: 5
   });
   expect(items.map(v => v.id)).toStrictEqual([1, 3, 2, 4]);
+
+  // Hamming distance
+  items = await Item.findAll({
+    order: hammingDistance('binary_embedding', '101', sequelize),
+    limit: 5
+  });
+  expect(items.map(v => v.id)).toStrictEqual([2, 3, 1, 4]);
+
+  // Jaccard distance
+  items = await Item.findAll({
+    order: jaccardDistance('binary_embedding', '101', sequelize),
+    limit: 5
+  });
+  expect(items.map(v => v.id)).toStrictEqual([2, 3, 1, 4]);
 
   // bad value
   await Item.create({embedding: 'bad'}).catch(e => expect(e.message).toMatch('invalid input syntax for type vector'));